% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_python.R
\name{as_pandas_DataFrame}
\alias{as_pandas_DataFrame}
\title{Convert from R to Pandas DataFrame using feather}
\usage{
as_pandas_DataFrame(x, ...)
}
\arguments{
\item{x}{An R object to convert to Pandas DataFrame}

\item{...}{Currently unused}
}
\description{
Convert any object coercible to a data.frame to a Pandas DataFrame.
Conversion is done using \code{feather}, which might not be the most efficient
way, but the type conversion supported by \code{feather} is inherited here.
}
