% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_python.R
\name{as_pandas_Series}
\alias{as_pandas_Series}
\title{Convert from R to Pandas Series using feather}
\usage{
as_pandas_Series(x, ...)
}
\arguments{
\item{x}{A 1-D R object to convert to a Pandas Series}

\item{...}{Currently unused}

\item{index}{A bare column name to use as the Series index if the R object
to convert has 2 columns.}
}
\description{
Convert R objects that are coercible to data.frames to Pandas Series.
These R objects must either have only 1 column, or must have 2 columns where
1 of those columns is identified as the \code{index} for the Series.
}
